requirements.txt
numpy
scipy
websocket-client
request
pyaudio
RealtimeSTT
ollama
tqdm
faster-whisper
PyYAML
transformers